+++ pkgs/zebra_datawedge/lib/src/zebra/zebra_interface.dart
@@ -9,8 +9,8 @@ class ZebraInterface {
 
   static EventChannel? _eventChannel;
 
-  static Future<bool> get isSupported async {
-    return await _channel.invokeMethod<bool>('isSupported') == true;
+  static Future<bool> get isControllerSupported async {
+    return await _channel.invokeMethod<bool>('isControllerSupported') == true;
   }
 
   static Future<List<String>> get profiles async {